server/requirements.txt
fastapi>=0.89.1,<0.90.0
uvicorn>=0.20.0,<0.21.0
typing>=3.4.0,<4.5.0
python-multipart>=0.0.5,<0.0.6
psycopg2-binary>=2.9.6,<3.0.0
python-time>=0.3.0,<0.3.1
python-jose>=3.3.0,<3.4.0
passlib>=1.7.4,<1.8.0
pydantic>=1.10.4,<2.0.0
